Move pluginHandle implementations into its own cpp
